Fix bad type hint
Using np.int here was causing the generated C code to do all the logic
involved with the full python ints and seriously slowing down the
preprocessing code.

My test dataset was preprocessed with more than a 20x speedup.

+++ mini_ziphmm_cython_funcs.pyx
@@ -47,7 +47,7 @@ def _forward(
 
 def preprocess_raw_observations(
         np.ndarray[np.int32_t, ndim=1] obs,
-        np.int nsyms):
+        np.int32_t nsyms):
     cdef:
         np.int32_t counted_last_pair, p_prev, p, new_sym
         np.int32_t max_pair, max_pair_left, max_pair_right, max_pair_count
